functions/processor/requirements.txt
google-cloud-pubsub==2.23.0
google-cloud-firestore==2.18.0
google-cloud-logging==3.11.2